refactor(editor): extract keymap and completion building into build-keymap-and-completion.ts (issue #1903)
Extracts the 75-line keymap, update listener, and completion source building logic
from Editor.svelte's onMount hook into a new pure function buildKeymapAndCompletion()
in src/renderer/lib/editor/build-keymap-and-completion.ts.

This is the second seam toward the broader CodeMirror extraction (issue #1903):
- buildExtensions() extracted in PR #1957 (#1921): 196-line extensions array
- buildKeymapAndCompletion() extracted here: keymap (Tab/Enter/Alt-Enter), update
  listener (onContentChange/onCursorChange tracking), and completion setup (tag /
  link / type-create sources + autocompletion configuration).

Metrics:
- Editor.svelte: 1024 → 962 lines (−62 lines, 6.0% reduction)
- build-keymap-and-completion.ts: 126 lines (created, under 600-line threshold)
- Combined with #1921 extraction: 1208 → 962 (−246 lines, 20.3% reduction)
- File-size budget updated
- All 625 test files pass (6827 tests)

Changes:
- New file: src/renderer/lib/editor/build-keymap-and-completion.ts
- Exports buildKeymapAndCompletion(opts): Extension[]
- Captures closure-dependent state through options interface:
  - Callbacks: onSave, openQuickFix, onContentChange, onCursorChange
  - Completion props: getNotePaths, getSources, getAliases, resolveInlineTypeCreate
  - Accessors for mutable state: getIgnoreNextUpdate, setIgnoreNextUpdate
  - The tagCompletion function + plainText flag
- Editor.svelte refactored to call buildKeymapAndCompletion() instead of inline
  - Removed 75 lines of keymap/completion building code
  - Cleaned up unused imports (keymap, Prec, autocompletion, etc.)
  - Updated file-size budget from 1024 to 962
