+++ src/renderer/lib/components/Editor.svelte
@@ -3,16 +3,15 @@
   import EditorContextMenu from './EditorContextMenu.svelte';
   import QuickFixMenu, { type QuickFix } from './QuickFixMenu.svelte';
   import { installDismissOnClickOutside } from '../dismiss-menu';
-  import { EditorView, keymap } from '@codemirror/view';
-  import { EditorState, Prec, Compartment } from '@codemirror/state';
+  import { EditorView } from '@codemirror/view';
+  import { EditorState, Compartment } from '@codemirror/state';
   import { cmTheme, fontSizeTheme, hiddenLineNumbersTheme } from '../editor/editor-theme';
   import { getEditorSettings, saveEditorSettings, type EditorSettings } from '../editor/settings';
   import { indentUnit, foldEffect, unfoldEffect, foldedRanges } from '@codemirror/language';
   import { highlightWhitespace } from '@codemirror/view';
   import { openSearchPanel, setSearchQuery, SearchQuery } from '@codemirror/search';
-  import { autocompletion, acceptCompletion, type CompletionContext, type CompletionResult } from '@codemirror/autocomplete';
-  import { acceptCompletionEatTail, completionKeymapNoEnter } from '../editor/accept-completion-eat-tail';
-  import { historyField, indentWithTab } from '@codemirror/commands';
+  import { type CompletionContext, type CompletionResult } from '@codemirror/autocomplete';
+  import { historyField } from '@codemirror/commands';
   import { api } from '../ipc/client';
   import {
     uploadImage,
@@ -21,7 +20,6 @@
     type UploadResult,
   } from '../editor/image-upload';
   import { sortLines } from '../editor/commands';
-  import { resolveKeyBindings } from '../editor/command-registry';
   import { toggleEditorDictation } from '../editor/dictation';
   import { findLinkAt, type LinkRange } from '../editor/link-decorations';
   import { brokenNoteLinkAt } from '../editor/broken-link-decorations';
@@ -31,8 +29,6 @@
     resolveBookmarkOffsets,
     type BookmarkRef,
   } from '../editor/bookmark-gutter';
-  import { linkCompletionSource } from '../editor/link-autocomplete';
-  import { typeCreateCompletionSource } from '../editor/type-create-autocomplete';
   import type { TypeInfo } from '../../../shared/objects/type-def';
   import { planBlockLink } from '../editor/block-link';
   import { toHistorySnapshot, canRestoreHistory } from '../editor/history-snapshot';
@@ -42,6 +38,11 @@
   import { extractClaimUri } from '../../../shared/refactor/find-arguments';
   import type { EditorContextMenuState, EditorMenuOps } from '../editor/context-menu-ops';
   import { buildExtensions } from '../editor/build-extensions';
+  import { buildKeymapAndCompletion } from '../editor/build-keymap-and-completion';
+
+  import { getToolInfosByCategory } from '../tools/tool-registry';
+  import { isSourceScoped } from '../../../shared/tools/types';
+  import { groupToolsByGroup } from '../../../shared/tools/grouping';
 
   export interface CursorInfo {
     line: number;
@@ -49,10 +50,6 @@
     selectionLength: number;
     wordCount: number;
   }
-
-  import { getToolInfosByCategory } from '../tools/tool-registry';
-  import { isSourceScoped } from '../../../shared/tools/types';
-  import { groupToolsByGroup } from '../../../shared/tools/grouping';
 
   interface Props {
     /**
@@ -682,82 +679,23 @@
   }
 
   onMount(() => {
-    const resolved = resolveKeyBindings();
-    const appKeymap = Prec.highest(keymap.of([
-      { key: 'Mod-s', run: () => { onSave(); return true; } },
-      { key: 'Alt-Enter', run: openQuickFix },
-      // Tab accepts the active completion, else indents (Shift-Tab outdents).
-      // acceptCompletion returns false with no popup open, and same-key bindings
-      // run in array order, so the fall-through lands on indentWithTab. Both
-      // no-op while tab-focus mode is on (editor.toggleTabFocusMode, Ctrl-m),
-      // which is what still lets keyboard users move focus out of the editor.
-      { key: 'Tab', run: acceptCompletion },
-      indentWithTab,
-      // Enter accepts the active completion and eats the half-typed word tail
-      // (#206) — only word chars, so `[[No|te]]` → `[[Notebook]]` keeps its
-      // brackets. Returns false with no popup open, so Enter stays a newline /
-      // list-continuation. completionKeymapNoEnter restores arrow-nav and
-      // Escape that defaultKeymap:false (below) removes.
-      { key: 'Enter', run: acceptCompletionEatTail },
-      ...completionKeymapNoEnter,
-      ...resolved.map(({ key: k, command: run }) => ({ key: k, run })),
-    ]));
-
-    const updateListener = EditorView.updateListener.of((update) => {
-      if (update.docChanged && !ignoreNextUpdate) {
-        onContentChange(update.state.doc.toString());
-      }
-      ignoreNextUpdate = false;
-
-      if (update.selectionSet || update.docChanged) {
-        const { state } = update;
-        const pos = state.selection.main.head;
-        const line = state.doc.lineAt(pos);
-        const sel = state.selection.main;
-        const docText = state.doc.toString();
-        onCursorChange?.({
-          line: line.number,
-          column: pos - line.from + 1,
-          selectionLength: Math.abs(sel.to - sel.from),
-          wordCount: docText.trim() ? docText.trim().split(/\s+/).length : 0,
-        });
-      }
-    });
-
-    const linkCompletion = linkCompletionSource({
-      getNotePaths: () => getNotePaths?.() ?? [],
-      getSources: () => getSources?.() ?? [],
-      getAliases: () => getAliases?.() ?? [],
-      readNote: (p) => api.notebase.readFile(p),
-    });
-
-    // Inline `/book` typed creation (#1065). The apply removes the `/partial`,
-    // then the host prompt+create runs async and its target is linked at the
-    // same spot — type, template, and link in one gesture.
-    const typeCreateCompletion = typeCreateCompletionSource({
-      loadTypes: () => api.types.list().then((c) => c.types),
-      onPick: (type, view, from, to) => {
-        view.dispatch({ changes: { from, to, insert: '' } });
-        void resolveInlineTypeCreate?.(type).then((target) => {
-          if (!target || !view.dom.isConnected) return;
-          const link = `[[${target}]]`;
-          const at = Math.min(from, view.state.doc.length);
-          view.dispatch({ changes: { from: at, insert: link }, selection: { anchor: at + link.length } });
-        });
-      },
-    });
-
-    const completion = autocompletion({
-      override: [tagCompletion, linkCompletion, typeCreateCompletion],
-      activateOnTyping: true,
-      closeOnBlur: true,
-      // Our Enter binding (acceptCompletionEatTail) owns accept-on-Enter; the
-      // built-in one would win the tie otherwise (#206).
-      defaultKeymap: false,
+    const keymapAndCompletionExtensions = buildKeymapAndCompletion({
+      onSave,
+      openQuickFix,
+      onContentChange,
+      onCursorChange,
+      getNotePaths,
+      getSources,
+      getAliases,
+      tagCompletion,
+      resolveInlineTypeCreate,
+      plainText,
+      getIgnoreNextUpdate: () => ignoreNextUpdate,
+      setIgnoreNextUpdate: (value) => { ignoreNextUpdate = value; },
     });
 
     // Wiki-link / tag autocomplete is markdown-only (#1130).
-    const allExtensions = [...extensions, appKeymap, updateListener, ...(plainText ? [] : [completion])];
+    const allExtensions = [...extensions, ...keymapAndCompletionExtensions];
 
     // When the caller passes a history snapshot AND its serialised doc
     // still matches the current content, restore the undo/redo stacks
