+++ tests/architecture/file-size-budgets.test.ts
@@ -60,7 +60,7 @@ const BUDGETS: Record<string, number> = {
   'src/renderer/lib/components/SourcesPanel.svelte': 1297,
   'src/renderer/lib/ipc/client.ts': 1239,
   'src/renderer/lib/stores/conversations.svelte.ts': 1212,
-  'src/renderer/lib/components/Editor.svelte': 1024,
+  'src/renderer/lib/components/Editor.svelte': 962,
   'src/renderer/lib/stores/editor.svelte.ts': 1183,
     'src/renderer/lib/components/right-sidebar/PropertiesPanel.svelte': 1156,
   'src/main/menu.ts': 1072,
